Added settings to increase performance on IE 11
On IE 11, adding DOM elements and forcing calculations of offsetWidth
causes lag. In complex applications, this can have a negative
performance impact.

I've added two settings values to account for this:

Settings.removeFromDOM (true by default)

Settings this to true, just hides the progress div rather than removing
it from the DOM. Calling Nprogress.Start shows it again. This has a
significant performance benefit in IE 11 as it minimizes calculations
for layout.

Settings.forceRedraw (true by default)

This removes the forced redraw from the progress bar as it is not needed
in all cases, and has a performance impact.